ceph: existence check before pool delete
Closes #3775.

+++ lxd/storage_ceph.go
@@ -270,11 +270,10 @@ func (s *storageCeph) StoragePoolDelete() error {
 		s.pool.Name, s.ClusterName)
 
 	// test if pool exists
-	if !cephOSDPoolExists(s.ClusterName, s.OSDPoolName, s.UserName) {
-		msg := fmt.Sprintf(`CEPH osd storage pool "%s" does not exist `+
+	poolExists := cephOSDPoolExists(s.ClusterName, s.OSDPoolName, s.UserName)
+	if !poolExists {
+		logger.Warnf(`CEPH osd storage pool "%s" does not exist `+
 			`in cluster "%s"`, s.OSDPoolName, s.ClusterName)
-		logger.Errorf(msg)
-		return fmt.Errorf(msg)
 	}
 
 	// Check whether we own the pool and only remove in this case.
@@ -285,13 +284,15 @@ func (s *storageCeph) StoragePoolDelete() error {
 			s.OSDPoolName, s.ClusterName)
 
 		// Delete the osd pool.
-		err := cephOSDPoolDestroy(s.ClusterName, s.OSDPoolName,
-			s.UserName)
-		if err != nil {
-			logger.Errorf(`Failed to delete CEPH OSD storage pool `+
-				`"%s" in cluster "%s": %s`, s.pool.Name,
-				s.ClusterName, err)
-			return err
+		if poolExists {
+			err := cephOSDPoolDestroy(s.ClusterName, s.OSDPoolName,
+				s.UserName)
+			if err != nil {
+				logger.Errorf(`Failed to delete CEPH OSD storage pool `+
+					`"%s" in cluster "%s": %s`, s.pool.Name,
+					s.ClusterName, err)
+				return err
+			}
 		}
 		logger.Debugf(`Deleted CEPH OSD storage pool "%s" in cluster "%s"`,
 			s.pool.Name, s.ClusterName)
@@ -299,15 +300,17 @@ func (s *storageCeph) StoragePoolDelete() error {
 
 	// Delete the mountpoint for the storage pool.
 	poolMntPoint := getStoragePoolMountPoint(s.pool.Name)
-	err := os.RemoveAll(poolMntPoint)
-	if err != nil {
-		logger.Errorf(`Failed to delete mountpoint "%s" for CEPH osd `+
-			`storage pool "%s" in cluster "%s": %s`, poolMntPoint,
-			s.OSDPoolName, s.ClusterName, err)
-		return err
+	if shared.PathExists(poolMntPoint) {
+		err := os.RemoveAll(poolMntPoint)
+		if err != nil {
+			logger.Errorf(`Failed to delete mountpoint "%s" for CEPH osd `+
+				`storage pool "%s" in cluster "%s": %s`, poolMntPoint,
+				s.OSDPoolName, s.ClusterName, err)
+			return err
+		}
+		logger.Debugf(`Deleted mountpoint "%s" for CEPH osd storage pool "%s" `+
+			`in cluster "%s"`, poolMntPoint, s.OSDPoolName, s.ClusterName)
 	}
-	logger.Debugf(`Deleted mountpoint "%s" for CEPH osd storage pool "%s" `+
-		`in cluster "%s"`, poolMntPoint, s.OSDPoolName, s.ClusterName)
 
 	logger.Infof(`Deleted CEPH OSD storage pool "%s" in cluster "%s"`,
 		s.pool.Name, s.ClusterName)
